+++ src/frontend/src/components/FKeyBar.jsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import styles from '../styles/FKeyBar.module.css'
 
-export default function FKeyBar({ onCopy, onMove, onNewDir, onDelete, onRename, onHelp }) {
+export default function FKeyBar({ onCopy, onMove, onNewDir, onDelete, onRename, onHelp, onView }) {
   const keys = [
     { num: 'F1', label: 'Help',   action: onHelp,   active: !!onHelp },
     { num: 'F2', label: 'Rename', action: onRename, active: !!onRename },
-    { num: 'F3', label: 'View',   action: null,     active: false },
+    { num: 'F3', label: 'View',   action: onView,   active: !!onView },
     { num: 'F4', label: 'Edit',   action: null,     active: false },
     { num: 'F5', label: 'Copy',   action: onCopy,   active: !!onCopy },
     { num: 'F6', label: 'Move',   action: onMove,   active: !!onMove },